refactor: move graphql-subscription
move the files from the subscription module into execution, in preparation for code integration
