+++ src/execution/index.ts
@@ -13,4 +13,8 @@ export type {
   FormattedExecutionResult,
 } from './execute';
 
+export { subscribe, createSourceEventStream } from './subscribe';
+
+export type { SubscriptionArgs } from './subscribe';
+
 export { getDirectiveValues } from './values';
